Add register rules and modularize app
