+++ controllers/.env.example
@@ -1,0 +1,2 @@
+# The database path including the file extension
+DATABASE_PATH=
