+++ templates/login.html
@@ -5,10 +5,10 @@ Login
 {% endblock %}
 
 {% block main %}
-<form action="{{ url_for('login') }}" method="post">
-    <input required type="text" name="username">
-    <input required type="password" name="password">
+<form action="{{ url_for('user.login') }}" method="post">
+    <input required type="text" name="username" placeholder="username">
+    <input required type="password" name="password" placeholder="password">
     <button type="submit">Enviar</button>
 </form>
-<button><a href="{{ url_for('register') }}">Create Account</a></button>
+<button><a href="{{ url_for('user.register') }}">Create Account</a></button>
 {% endblock %}
